separate modelprocessor
